Update README links, refactor App router basename, implement useMedia hook for media fetching in Home component
