+++ vite.config.js
@@ -4,5 +4,5 @@ import react from '@vitejs/plugin-react-swc';
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
-  base: '/~quangth/week4-react-hooks/',
+  base: '/~quangth/week5-react-custom-hooks/',
 });
